Switched color to integer field

+++ Assets/Swagger/src/main/CsharpDotNet2/IO/Swagger/Model/KeyboardParam.cs
@@ -18,7 +18,7 @@ namespace IO.Swagger.Model {
     /// <value>Color value in BGR format</value>
     [DataMember(Name="color", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "color")]
-    public double? Color { get; set; }
+    public int? Color { get; set; }
 
 
     /// <summary>
